Randomized Quick Sort More efficient version
The code for randomized quicksort uses a rand() function which generates a random number
between a given range. We choose a random index and swap the element at that index with the
last element. After this step, we make the new last element as pivot. Generated Output for the code is as follows:-

Array before sorting:-
10 7 8 9 5 1 5
Array after sorting:-
1 5 5 7 8 9 10