Gaussian random numbers
The program finds the first seed that is greater or equal to K where each of N Gaussian numbers is lesser than or equal M.

The input contains numbers K, N, M in a single line.

You should output the seed.

Hint

You have to check all n iterations of Gaussian generated numbers to be less than M. If that is true, then escape loop and print the seed. If any of the generated numbers is greater than or equal to M, then you have to test out the next seed and go to the loop all over again.
Don't get too upset if you still don't understand this problem. Skip it or google for more information.